Safari-609.1.11.1/LayoutTests/js/dom/Promise-resolve-state.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>
description('Test Promise.');

window.jsTestIsAsync = true;

var resolve;
var promise = new Promise(function(r) { resolve = r; });
var promiseState = 'pending';
var promiseResult = undefined;
promise.then(function(result) {
  promiseState = 'fulfilled';
  promiseResult = result;
}, function(result) {
  promiseState = 'rejected';
  promiseResult = result;
});

shouldBeEqualToString('promiseState', 'pending');

resolve('hello');

shouldBeEqualToString('promiseState', 'pending');

promise.then(function() {
  shouldBeEqualToString('promiseState', 'fulfilled');
  shouldBeEqualToString('promiseResult', 'hello');
  finishJSTest();
}, function() {
  testFailed('promise is rejected.');
  finishJSTest();
});

</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
